refactor(cli): move logs command to app/cli/logs.py
Moves logs_cmd, _follow_logs, and _LOG_SHOW_CHOICES out of app/main.py
into app/cli/logs.py. Drops orphaned re-imports (time, _resolve_flow_id,
_render_log_entry, _truncated_section, _build_failure_body,
_TASK_STATUS_GLYPHS, FlowStatus) from app/main.py — those names had no
remaining users.
